@@ -1,6 +1,6 @@
 buildscript {
 
-    ext.kotlin_version = '1.2.71'
+    ext.kotlin_version = '1.8.0'
     ext.support = "28.0.0"
 
     repositories {
@@ -14,9 +14,9 @@ buildscript {
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.1'
+        classpath 'com.android.tools.build:gradle:7.4.2'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'io.fabric.tools:gradle:1.26.1'
+        classpath 'io.fabric.tools:gradle:1.28.0'
     }
 }
 
