@@ -4,28 +4,32 @@ import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
-import kotlinx.android.synthetic.main.toolbar.*
+import android.support.v7.widget.Toolbar
 import melerospaw.deudapp.R
 import melerospaw.deudapp.iu.fragments.FragmentPreferencias
 
 
-class ActivityPreferencias: AppCompatActivity() {
+class ActivityPreferencias : AppCompatActivity() {
 
     companion object {
 
-        @JvmStatic fun start(context: Context) {
+        @JvmStatic
+        fun start(context: Context) {
             context.startActivity(Intent(context, ActivityPreferencias::class.java))
         }
     }
 
+    private var toolbar: Toolbar? = null
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_preferencias)
+        toolbar = findViewById(R.id.toolbar)
         setSupportActionBar(toolbar)
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
 
         supportFragmentManager.beginTransaction()
-                .replace(R.id.fl_root_currency_prefs, FragmentPreferencias(), FragmentPreferencias.TAG)
-                .commit()
+            .replace(R.id.fl_root_currency_prefs, FragmentPreferencias(), FragmentPreferencias.TAG)
+            .commit()
     }
 }
