@@ -5,15 +5,14 @@ import android.support.v4.app.DialogFragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import kotlinx.android.synthetic.main.dialog_time_picker.*
+import android.widget.DatePicker
+import android.widget.TextView
 import melerospaw.deudapp.R
 import melerospaw.deudapp.utils.inflate
-import java.util.*
+import java.util.Calendar
+import java.util.Date
 
 class DialogFechaDeuda : DialogFragment() {
-
-    private lateinit var fecha: Date
-    var positiveCallback: PositiveCallback? = null
 
     companion object {
 
@@ -31,35 +30,58 @@ class DialogFechaDeuda : DialogFragment() {
         }
     }
 
+    private var dpFecha: DatePicker? = null
+    private var tvCambiar: TextView? = null
+    private var tvCancelar: TextView? = null
+
+    private lateinit var fecha: Date
+    var positiveCallback: PositiveCallback? = null
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         fecha = Date(arguments!!.getLong(BUNDLE_FECHA))
     }
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? =
-            inflate(inflater, R.layout.dialog_time_picker, container)
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View? = inflate(inflater, R.layout.dialog_time_picker, container).also {
+        bindViews(it)
+        bindViews(it)
+        bindViews(it)
+    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        val cal = Calendar.getInstance()
-        cal.time = fecha
-        dpFecha.init(cal.get(Calendar.YEAR), cal.get(Calendar.MONTH), cal.get(Calendar.DAY_OF_MONTH), null)
+        initDate()
+        tvCambiar?.setOnClickListener { guardar() }
+        tvCancelar?.setOnClickListener { dismiss() }
+    }
 
-        tvCambiar.setOnClickListener {
-            guardar()
+    private fun bindViews(view: View) {
+        with(view) {
+            dpFecha = findViewById(R.id.dp_fecha)
+            tvCambiar = findViewById(R.id.tv_cambiar)
+            tvCancelar = findViewById(R.id.tv_cancelar)
         }
-        tvCancelar.setOnClickListener {
-            dismiss()
+    }
+
+    private fun initDate() {
+        dpFecha?.let {
+            val cal = Calendar.getInstance().apply { time = fecha }
+            it.init(cal.get(Calendar.YEAR), cal.get(Calendar.MONTH), cal.get(Calendar.DAY_OF_MONTH), null)
         }
     }
 
     private fun guardar() {
-        val cal = Calendar.getInstance()
-        cal.set(Calendar.YEAR, dpFecha.year)
-        cal.set(Calendar.MONTH, dpFecha.month)
-        cal.set(Calendar.DAY_OF_MONTH, dpFecha.dayOfMonth)
-        positiveCallback?.guardarFecha(cal.time)
-        dismiss()
+        dpFecha?.run {
+            val cal = Calendar.getInstance()
+            cal.set(Calendar.YEAR, year)
+            cal.set(Calendar.MONTH, month)
+            cal.set(Calendar.DAY_OF_MONTH, dayOfMonth)
+            positiveCallback?.guardarFecha(cal.time)
+            dismiss()
+        }
     }
 
     interface PositiveCallback {
